Use BACNET_OBJECT_TYPE for all object type variables.

+++ src/bacnet/getevent.c
@@ -56,7 +56,7 @@ int getevent_encode_apdu(uint8_t *apdu,
         /* encode optional parameter */
         if (lastReceivedObjectIdentifier) {
             len = encode_context_object_id(&apdu[apdu_len], 0,
-                (int)lastReceivedObjectIdentifier->type,
+                lastReceivedObjectIdentifier->type,
                 lastReceivedObjectIdentifier->instance);
             apdu_len += len;
         }
@@ -117,7 +117,7 @@ int getevent_ack_encode_apdu_data(uint8_t *apdu,
         while (event_data) {
             /* Tag 0: objectIdentifier */
             apdu_len += encode_context_object_id(&apdu[apdu_len], 0,
-                (int)event_data->objectIdentifier.type,
+                event_data->objectIdentifier.type,
                 event_data->objectIdentifier.instance);
             /* Tag 1: eventState */
             apdu_len += encode_context_enumerated(
